Big restructure of common functions + f32-ization

+++ src/common/structs/mod.rs
@@ -150,9 +150,9 @@ pub struct Replay {
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct ReplayNote {
 	/// The position of the note inside the chart, in seconds
-	pub time: f64,
+	pub time: f32,
 	/// The offset that the note was hit with, in seconds. A 50ms early hit would be `-0.05`
-	pub deviation: f64,
+	pub deviation: f32,
 	/// The position of the ntoe inside the chart, in ticks (192nds)
 	pub tick: Option<u32>,
 	/// The lane/column that this note appears on. 0-3 for 4k, 0-5 for 6k
